refactor(Step): move package
