+++ src/main/java/nextstep/ladder/domain/step/MovableStrategy.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.step;
 
 @FunctionalInterface
 public interface MovableStrategy {
