+++ src/main/java/nextstep/ladder/domain/step/RandomStrategy.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.step;
 
 import java.util.Random;
