+++ src/main/java/nextstep/ladder/domain/step/Step.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.step;
 
 public class Step {
     private final int position;
