+++ src/main/java/nextstep/ladder/domain/step/Steps.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.step;
 
 import java.util.ArrayList;
 import java.util.List;
